Fix compatibility with pytest 3.8.0
As discussed in https://github.com/pytest-dev/pytest/issues/3973, pytest 3.8.0
introduced some code that ends up checking the `cwd` by accident. One of
the tests in your test suites changes the `cwd` to a temporary directory, and
this blows up some tests after it.

This change uses the `monkeypatch` fixture to safe cwd-switching, by
restoring it to the original value at the end of the test.